pyproject.toml
[tool.poetry]
name = "simulation-bridge"
version = "0.1.0"
description = "Simulation Bridge"
authors = ["Marco Melloni <291358@studenti.unimore.it>"]
license = "ICAPL"
readme = "README.md"

packages = [{ include = "simulation_bridge"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = ">=2.1.0,<3.0.0"
pyyaml = ">=6.0.2,<7.0.0"
pika = "^1.3.2"
scipy = "^1.15.2"
matplotlib = "^3.10.1"
python-json-logger = "^3.3.0"
pydantic = "^2.11.3"
psutil = "^7.0.0"
colorlog = "^6.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
flake8 = "^7.2.0"
types-pyyaml = "^6.0.12.20250402"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
simulation-bridge = "simulation_bridge.src.main:main"